# contracts/openapi/risk-aiops-v1.yaml
openapi: 3.1.0
info:
  title: Financial Trading Risk AIOps API
  version: 1.0.0
  description: Read-only diagnosis plus governed report draft/decision APIs over fictional reference data.
servers:
  - url: http://localhost:3000
security:
  - oidc: []
paths:
  /api/v1/diagnoses:
    post:
      operationId: createDiagnosis
      security: [{ oidc: [diagnosis:write] }]
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, minLength: 8, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDiagnosis' }
      responses:
        '202': { description: Accepted }
        '409': { description: Idempotency key conflict }
  /api/v1/diagnoses/{id}:
    get:
      operationId: getDiagnosis
      security: [{ oidc: [diagnosis:read] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: Diagnosis state }
        '404': { description: Not found }
  /api/v1/diagnoses/{id}/events:
    get:
      operationId: streamDiagnosisEvents
      security: [{ oidc: [diagnosis:read] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Server-sent state transitions
          content: { text/event-stream: { schema: { type: string } } }
  /api/v1/reports:
    get:
      operationId: listReports
      security: [{ oidc: [report:read] }]
      responses: { '200': { description: Governed report list } }
    post:
      operationId: createReportDraft
      security: [{ oidc: [report:write] }]
      responses: { '201': { description: Draft created } }
  /api/v1/reports/{id}:
    get:
      operationId: getReport
      security: [{ oidc: [report:read] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses: { '200': { description: Report } }
  /api/v1/reports/{id}/decisions:
    post:
      operationId: decideReport
      description: Creator cannot approve. If-Match is required for optimistic locking.
      security: [{ oidc: [report:approve] }]
      parameters:
        - $ref: '#/components/parameters/Id'
        - in: header
          name: If-Match
          required: true
          schema: { type: string }
      responses:
        '200': { description: Decision recorded }
        '409': { description: Maker-checker or version conflict }
        '428': { description: If-Match required }
  /api/v1/audit-events:
    get:
      operationId: listAuditEvents
      security: [{ oidc: [audit:read] }]
      responses: { '200': { description: Authorized append-only audit events } }
  /api/v1/risk-snapshots:
    get:
      operationId: getRiskSnapshot
      security: [{ oidc: [risk:read] }]
      parameters:
        - in: query
          name: accountId
          required: true
          schema: { type: string, pattern: '^[A-Z0-9_-]{1,64}$' }
        - in: query
          name: asOf
          schema: { type: string, format: date-time }
      responses: { '200': { description: Read-only risk snapshot envelope } }
  /api/v1/reconciliation-breaks:
    get:
      operationId: getReconciliationBreaks
      security: [{ oidc: [reconciliation:read] }]
      parameters:
        - in: query
          name: accountId
          required: true
          schema: { type: string, pattern: '^[A-Z0-9_-]{1,64}$' }
        - in: query
          name: tradeDate
          required: true
          schema: { type: string, format: date }
      responses: { '200': { description: Read-only OMS/broker reconciliation breaks } }
  /api/v1/evidence/{id}:
    get:
      operationId: getEvidence
      security: [{ oidc: [evidence:read] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, minLength: 4, maxLength: 128 }
      responses:
        '200': { description: Stored evidence document }
        '404': { description: Not found }
  /api/v1/incidents:
    get:
      operationId: listIncidents
      security: [{ oidc: [incident:read] }]
      parameters:
        - in: query
          name: accountId
          schema: { type: string, pattern: '^[A-Z0-9_-]{1,64}$' }
      responses: { '200': { description: Authorized incidents } }
    post:
      operationId: createIncident
      security: [{ oidc: [incident:write] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateIncident' }
      responses: { '201': { description: Incident created } }
  /api/v1/incidents/{id}/ack:
    post:
      operationId: acknowledgeIncident
      security: [{ oidc: [incident:write] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses: { '200': { description: Incident acknowledged } }
  /api/v1/incidents/{id}/close:
    post:
      operationId: closeIncident
      security: [{ oidc: [incident:write] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CloseIncident' }
      responses: { '200': { description: Incident closed } }
  /internal/v1/events:
    post:
      operationId: ingestSourceEvent
      description: Internal read-only data ingress. Production requires source mTLS and ingest:write scope.
      security: [{ oidc: [ingest:write] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EventEnvelope' }
      responses:
        '202': { description: Event accepted for replay processing }
        '400': { description: Invalid event envelope }
        '403': { description: Source mTLS required }
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: http://localhost:8081/realms/risk-aiops/.well-known/openid-configuration
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
  schemas:
    Decimal:
      type: string
      pattern: '^-?(0|[1-9][0-9]*)(\.[0-9]{1,10})?$'
      examples: ['123456.7800000000']
    CreateDiagnosis:
      type: object
      additionalProperties: false
      required: [accountId, tradeDate]
      properties:
        accountId: { type: string, pattern: '^[A-Z0-9_-]{1,64}$' }
        tradeDate: { type: string, format: date }
    CreateIncident:
      type: object
      additionalProperties: false
      required: [severity, title]
      properties:
        accountId: { type: string, pattern: '^[A-Z0-9_-]{1,64}$' }
        severity: { type: string, enum: [INFO, WARNING, CRITICAL] }
        title: { type: string, minLength: 1, maxLength: 256 }
    CloseIncident:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason: { type: string, minLength: 1, maxLength: 2048 }
    EventEnvelope:
      type: object
      additionalProperties: false
      required: [eventId, schemaVersion, eventType, source, sourceTimestamp, publishedAt, partitionKey, payload]
      properties:
        eventId: { type: string, format: uuid }
        schemaVersion: { type: string, enum: ['1.0'] }
        eventType:
          type: string
          enum: [market.snapshot.v1, position.snapshot.v1, order.lifecycle.v1, execution.report.v1, risk.limit.v1, system.health.v1, incident.signal.v1]
        source: { type: string, minLength: 1, maxLength: 128 }
        sourceTimestamp: { type: string, format: date-time }
        publishedAt: { type: string, format: date-time }
        partitionKey: { type: string, minLength: 1, maxLength: 256 }
        payload: { type: object }